requirements.txt
pandas==2.2.3
usd-core==24.8
pydantic==2.6.4
python-dotenv==1.0.1
llama-index==0.11.19
openai==1.12.0
llama-index-core==0.10.10
typing-extensions==4.10.0